[refactor] 각 템플릿 페이지 폴더 내로 이동
